Provisioning routing fixed
