+++ src/router/provisionWorkflow.ts
@@ -8,16 +8,110 @@ import AddCSPAdmin from "@/portfolios/provisioning/AddCSPAdmin.vue";
 import ReadyToProvision from "@/portfolios/provisioning/ReadyToProvision.vue";
 import AddToExistingPortfolio from "@/portfolios/provisioning/AddToExistingPortfolio.vue"
 import ProvisioningIssue from "@/portfolios/provisioning/ProvisioningIssue.vue"
+import { StepRouteResolver } from "@/store/steps/types";
+import PortfolioStore from "@/store/portfolio";
+import AcquisitionPackageSummary from "@/store/acquisitionPackageSummary";
+import PortfolioSummary from "@/store/portfolioSummary";
 
-const GeneratedFromPackageRouteResolver = (): string => {
-  return "";
+const provFromMeatball = (): boolean => {
+  return PortfolioStore.provisioningFromMeatball
 }
-const PortfolioDetailsRouteResolver = (): string => {
-  return "";
+const cspHasILs = (): boolean => {
+  return PortfolioStore.CSPHasImpactLevels
 }
-const AddToExistingPortfolioResolver = (): string => {
-  return "";
+const taskOrderHasUnclass = (): boolean => {
+  return PortfolioStore.doesTaskOrderHaveUnclassified
 }
+const userHasActivePortfolios = (): boolean => {
+  return PortfolioSummary.hasActivePortfolios
+}
+
+export const AddToExistingPortfolioResolver = (current: string): string => {
+  const hasActivePortfolios: boolean = userHasActivePortfolios()
+  // moving backward
+  if (
+    current === provWorkflowRouteNames.GeneratedFromPackage ||
+		current === provWorkflowRouteNames.PortfolioDetails
+  ) {
+    return hasActivePortfolios
+      ? provWorkflowRouteNames.AddToExistingPortfolio
+      : provWorkflowRouteNames.AwardedTaskOrder
+  }
+
+  // moving forward
+  if (provFromMeatball()) {
+    return taskOrderHasUnclass() && cspHasILs()
+      ? provWorkflowRouteNames.PortfolioDetails
+      : provWorkflowRouteNames.AddCSPAdmin
+  }
+
+  if (hasActivePortfolios) return provWorkflowRouteNames.AddToExistingPortfolio
+
+  return GeneratedFromPackageRouteResolver(current)
+}
+
+export const GeneratedFromPackageRouteResolver = (current: string): string => {
+  const packageCount = AcquisitionPackageSummary.packagesWaitingForTaskOrderCount;
+  const acqPkgSysId = PortfolioStore.getSelectedAcquisitionPackageSysId
+  const showPackageSelection = PortfolioStore.showTOPackageSelection
+
+  if (packageCount && (!acqPkgSysId || showPackageSelection)) {
+    return provWorkflowRouteNames.GeneratedFromPackage
+  }
+
+  if (current !== provWorkflowRouteNames.PortfolioDetails && acqPkgSysId && !cspHasILs()) {
+    return provWorkflowRouteNames.AddCSPAdmin
+  }
+
+  if (current === provWorkflowRouteNames.PortfolioDetails) {
+    if (provFromMeatball()) return provWorkflowRouteNames.AwardedTaskOrder;
+    return userHasActivePortfolios()
+      ? provWorkflowRouteNames.AddToExistingPortfolio
+      : provWorkflowRouteNames.AwardedTaskOrder
+  }
+
+  if (!acqPkgSysId && current === provWorkflowRouteNames.AwardedTaskOrder) {
+    return provWorkflowRouteNames.PortfolioDetails;
+  }
+
+  return taskOrderHasUnclass() && cspHasILs()
+    ? provWorkflowRouteNames.PortfolioDetails
+    : provWorkflowRouteNames.AddCSPAdmin
+}
+
+export const PortfolioDetailsRouteResolver = (current: string): string => {
+  if (current === provWorkflowRouteNames.AddCSPAdmin && provFromMeatball()) {
+    return taskOrderHasUnclass() && cspHasILs()
+      ? provWorkflowRouteNames.PortfolioDetails
+      : provWorkflowRouteNames.AwardedTaskOrder
+  }
+  const acqPkgSysId = PortfolioStore.getSelectedAcquisitionPackageSysId
+  if (!acqPkgSysId || (taskOrderHasUnclass() && cspHasILs())) {
+    return provWorkflowRouteNames.PortfolioDetails
+  }
+  if (
+    current === provWorkflowRouteNames.AddCSPAdmin &&
+		acqPkgSysId &&
+		!taskOrderHasUnclass()
+  ) {
+    return provWorkflowRouteNames.GeneratedFromPackage
+  }
+  if (
+    current === provWorkflowRouteNames.AddCSPAdmin &&
+		!acqPkgSysId &&
+		!taskOrderHasUnclass()
+  ) {
+    return userHasActivePortfolios()
+      ? provWorkflowRouteNames.AddToExistingPortfolio
+      : provWorkflowRouteNames.AwardedTaskOrder
+  }
+  // eslint-disable-next-line max-len
+  return current === provWorkflowRouteNames.GeneratedFromPackage ||
+		provWorkflowRouteNames.AddToExistingPortfolio
+    ? provWorkflowRouteNames.AddCSPAdmin
+    : provWorkflowRouteNames.GeneratedFromPackage
+}
+
 
 export const provWorkflowRouteNames = {
   ProvisioningIndex: "Provisioning_Index",
@@ -116,6 +210,19 @@ export const provisionWorkFlowRoutes: Array<StepperRouteConfig> = [
 
 ];
 
+const provisioningRouteResolvers: Record<string, StepRouteResolver> = {
+  GeneratedFromPackageRouteResolver,
+  AddToExistingPortfolioResolver,
+  PortfolioDetailsRouteResolver,
+}
+
+export const InvokeATATRouteResolver = (
+  resolverName: string,
+  currentStep: string
+): string => provisioningRouteResolvers[resolverName](currentStep)
+
+
+
 /**
  * Helper method to convert Stepper Route
  * to Stepper Step
@@ -138,7 +245,7 @@ const mapStepRouteToStepperData = (
     altContinueAction,
   } = stepperRouteConfig;
   let {name} = stepperRouteConfig;
-  name = name || "";
+  name = name as string || "";
 
   const stepperStep: StepperStep = {
     stepNumber,
